feat: 유저, 모임 정보 수정때 기존 이미지 없으면 에러발생 수정

+++ src/main/java/com/example/moim/user/entity/User.java
@@ -105,7 +105,9 @@ public class User extends BaseEntity {
             this.phone = userUpdateInput.getPhone();
         }
         if (imgPath != null) {
-            new File(this.imgPath).delete();
+            if (this.imgPath != null) {
+                new File(this.imgPath).delete();
+            }
             this.imgPath = imgPath;
         }
         if (userUpdateInput.getGender() != null && !userUpdateInput.getGender().isBlank()) {
